requirements.txt
certifi==2020.6.20
chardet==3.0.4
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
Naked==0.1.31
netifaces==0.10.9
pycodestyle==2.6.0
pycryptodome==3.9.8
pyflakes==2.2.0
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
urllib3==1.25.9
zipp==3.1.0
